# apps/api/cmd/api/main.go
// x-web3 api entrypoint.
//
// 启动顺序：
//  1. 加载 config（fail-fast）
//  2. zap logger
//  3. PostgreSQL pool（pgx）
//  4. Redis client
//  5. Privy verifier（dev stub 可选）
//  6. Session store
//  7. RBAC engine
//  8. Audit writer
//  9. Catalog service（带缓存订阅）
// 10. HTTP router + handlers
// 11. graceful shutdown
package main

import (
	"context"
	"errors"
	"net/http"
	"os/signal"
	"strconv"
	"strings"
	"syscall"
	"time"

	"github.com/gin-gonic/gin"
	"github.com/jackc/pgx/v5/pgxpool"
	"github.com/redis/go-redis/v9"
	"go.uber.org/zap"

	"github.com/x-web3/api/internal/audit"
	"github.com/x-web3/api/internal/auth"
	"github.com/x-web3/api/internal/catalog"
	"github.com/x-web3/api/internal/comment"
	"github.com/x-web3/api/internal/config"
	"github.com/x-web3/api/internal/course"
	"github.com/x-web3/api/internal/handlers"
	"github.com/x-web3/api/internal/httpkit"
	"github.com/x-web3/api/internal/learning"
	"github.com/x-web3/api/internal/media"
	"github.com/x-web3/api/internal/objectstore"
	"github.com/x-web3/api/internal/rbac"
	"github.com/x-web3/api/internal/user"
	"github.com/x-web3/api/internal/wallet"
)

func main() {
	// 自动从 CWD 或祖先目录加载 .env；找不到也不报错（prod 走真 env）
	dotenv := config.LoadDotenv()
	cfg := config.MustLoad()

	logger, err := newLogger(cfg)
	if err != nil {
		panic(err)
	}
	defer logger.Sync() //nolint:errcheck

	ctx, stop := signal.NotifyContext(context.Background(), syscall.SIGINT, syscall.SIGTERM)
	defer stop()

	pool, err := pgxpool.New(ctx, cfg.DatabaseURL)
	if err != nil {
		logger.Fatal("pgx_connect_failed", zap.Error(err))
	}
	defer pool.Close()
	if err := pool.Ping(ctx); err != nil {
		logger.Fatal("pg_ping_failed", zap.Error(err))
	}

	redisOpts, err := redis.ParseURL(cfg.RedisURL)
	if err != nil {
		logger.Fatal("redis_parse_url", zap.Error(err))
	}
	rdb := redis.NewClient(redisOpts)
	defer rdb.Close()
	if err := rdb.Ping(ctx).Err(); err != nil {
		logger.Fatal("redis_ping_failed", zap.Error(err))
	}

	verifier, err := auth.NewVerifier(ctx, auth.Config{
		AppID:      cfg.PrivyAppID,
		JWKSURL:    cfg.PrivyJWKSURL,
		Audience:   cfg.PrivyAudience,
		DevStub:    cfg.PrivyDevStub,
		DevSubject: cfg.PrivyDevSubject,
	}, logger)
	if err != nil {
		logger.Fatal("privy_verifier_init", zap.Error(err))
	}

	sessionStore := auth.NewSessionStore(rdb, cfg.SessionSecret, cfg.SessionTTL)
	nonceStore := wallet.NewNonceStore(rdb, cfg.WalletNonceTTL)
	auditWriter := audit.NewWriter(pool, logger)
	rbacEngine := rbac.NewEngine(pool, logger)
	walletSvc := wallet.NewService(pool, nonceStore, cfg.APIDomain, auditWriter)

	// 业务子系统
	courseRepo := course.NewRepo(pool)
	catalogSvc := catalog.NewService(courseRepo, rdb)
	mediaRepo := media.NewRepo(pool)
	commentRepo := comment.NewRepo(pool)

	// Object store：MVP 用 dev fake；F07 替换为 AWS S3。
	objStore := objectstore.NewFakeStore()

	learningSvc := learning.NewService(pool, objStore)

	router := httpkit.NewRouter(logger, cfg.WebOrigin)
	v1 := router.Engine.Group("/api/v1")

	// Health
	router.Engine.GET("/healthz", func(c *gin.Context) {
		c.JSON(http.StatusOK, gin.H{"status": "ok"})
	})
	router.Engine.GET("/readyz", func(c *gin.Context) {
		if err := pool.Ping(c.Request.Context()); err != nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{"status": "db_unavailable"})
			return
		}
		if err := rdb.Ping(c.Request.Context()).Err(); err != nil {
			c.JSON(http.StatusServiceUnavailable, gin.H{"status": "redis_unavailable"})
			return
		}
		c.JSON(http.StatusOK, gin.H{"status": "ok"})
	})

	authH := handlers.NewAuthHandler(cfg, pool, verifier, sessionStore, auditWriter, logger)
	walletH := handlers.NewWalletHandler(cfg, pool, walletSvc, auditWriter, logger)
	meH := handlers.NewMeHandler(pool, auditWriter, logger, authH)
	courseH := handlers.NewCourseHandler(courseRepo, catalogSvc, auditWriter)
	mediaH := handlers.NewMediaHandler(mediaRepo, objStore, auditWriter, logger)
	learningH := handlers.NewLearningHandler(learningSvc, auditWriter, logger)
	commentH := handlers.NewCommentHandler(commentRepo, auditWriter, logger)

	authGroup := v1.Group("/auth")
	{
		authGroup.POST("/privy/session", httpkit.RateLimit(rdb, "login", cfg.LoginRateLimit, httpkit.ClientIPKey), httpkit.Wrap(authH.PostPrivySession))
		authGroup.POST("/session/refresh", auth.Middleware(verifier, sessionStore, pool), httpkit.Wrap(authH.RefreshSession))
		authGroup.DELETE("/session", httpkit.Wrap(authH.DeleteSession))
	}
	meGroup := v1.Group("/me")
	meGroup.Use(auth.Middleware(verifier, sessionStore, pool))
	{
		meGroup.GET("", httpkit.Wrap(meH.GetMe))
		walletLimit := httpkit.RateLimit(rdb, "wallet", cfg.WalletRateLimit, httpkit.UserIDKeyFunc)
		meGroup.POST("/wallets/nonce", walletLimit, httpkit.Wrap(walletH.IssueNonce))
		meGroup.POST("/wallets/link", walletLimit, httpkit.Wrap(walletH.Link))
		meGroup.DELETE("/wallets/:walletId", walletLimit, httpkit.Wrap(walletH.Unbind))
	}
	catalogGroup := v1.Group("/courses")
	{
		catalogGroup.GET("", httpkit.Wrap(courseH.List))
		catalogGroup.GET("/:id", httpkit.Wrap(courseH.Get))
		catalogGroup.GET("/:id/comments", httpkit.Wrap(commentH.GetCourseComments))
	}
	coursesAuthGroup := v1.Group("/courses")
	coursesAuthGroup.Use(auth.Middleware(verifier, sessionStore, pool))
	{
		coursesAuthGroup.POST("/:id/comments", httpkit.Wrap(commentH.PostCreate))
		coursesAuthGroup.DELETE("/comments/:id", httpkit.Wrap(commentH.DeleteMine))
	}
	teacherGroup := v1.Group("/teacher")
	teacherGroup.Use(auth.Middleware(verifier, sessionStore, pool))
	{
		teacherGroup.POST("/courses", rbacEngine.Middleware(user.PermCourseCreate), httpkit.Wrap(courseH.Create))
		teacherGroup.PUT("/courses/:id", rbacEngine.Middleware(user.PermCourseEdit), httpkit.Wrap(courseH.Update))
		teacherGroup.PUT("/courses/:id/curriculum", rbacEngine.Middleware(user.PermCourseEdit), httpkit.Wrap(courseH.ReplaceCurriculum))
		teacherGroup.POST("/courses/:id/submit", rbacEngine.Middleware(user.PermCourseEdit), httpkit.Wrap(courseH.Submit))
		teacherGroup.POST("/media/upload-intent", httpkit.Wrap(mediaH.PostUploadIntent))
		teacherGroup.POST("/media/:id/finalize", httpkit.Wrap(mediaH.PostFinalize))
		teacherGroup.GET("/media", httpkit.Wrap(mediaH.GetMine))
		teacherGroup.GET("/lessons/:id/preview", httpkit.Wrap(learningH.GetPreview))
	}
	courseAdminGroup := v1.Group("/admin/courses")
	courseAdminGroup.Use(auth.Middleware(verifier, sessionStore, pool), rbacEngine.Middleware(user.PermCourseApprove))
	{
		courseAdminGroup.POST("/:id/review", httpkit.Wrap(courseH.Review))
		courseAdminGroup.POST("/:id/archive", httpkit.Wrap(courseH.Archive))
	}
	commentsAdminGroup := v1.Group("/admin/comments")
	commentsAdminGroup.Use(auth.Middleware(verifier, sessionStore, pool), rbacEngine.Middleware(user.PermCommentModerate))
	{
		commentsAdminGroup.PATCH("/:id", httpkit.Wrap(commentH.PatchModerate))
	}
	learningGroup := v1.Group("/lessons")
	learningGroup.Use(auth.Middleware(verifier, sessionStore, pool))
	{
		learningGroup.GET("/:id/playback", httpkit.Wrap(learningH.GetPlayback))
	}

	adminGroup := v1.Group("/admin")
	adminGroup.Use(auth.Middleware(verifier, sessionStore, pool), rbacEngine.Middleware(user.PermSystemAdmin))
	adminGroup.GET("/ping", func(c *gin.Context) { c.JSON(http.StatusOK, gin.H{"status": "ok"}) })

	// 缓存失效订阅
	go func() {
		if err := catalogSvc.SubscribeInvalidate(ctx); err != nil && !errors.Is(err, context.Canceled) {
			logger.Warn("catalog_invalidate_subscriber_exit", zap.Error(err))
		}
	}()

	logger.Info("api_starting",
		zap.String("env", cfg.Env),
		zap.Int("port", cfg.APIPort),
		zap.Bool("privy_dev_stub", cfg.PrivyDevStub),
		zap.String("dotenv_path", dotenv.Path),
		zap.Bool("dotenv_loaded", dotenv.Loaded),
		zap.String("cwd", dotenv.CWD),
		zap.String("database", maskURL(cfg.DatabaseURL)),
		zap.String("redis", maskURL(cfg.RedisURL)),
		zap.String("object_store_region", objStore.Region()),
		zap.String("object_store_bucket", objStore.Bucket()),
	)
	if !dotenv.Loaded {
		logger.Warn("dotenv_not_loaded",
			zap.String("cwd", dotenv.CWD),
			zap.Strings("candidates", dotenv.Candidates),
			zap.Error(dotenv.Err),
		)
	}

	srv := &http.Server{
		Addr:              ":" + strconv.Itoa(cfg.APIPort),
		Handler:           router.Engine,
		ReadHeaderTimeout: 10 * time.Second,
	}
	go func() {
		if err := srv.ListenAndServe(); err != nil && !errors.Is(err, http.ErrServerClosed) {
			logger.Fatal("listen_failed", zap.Error(err))
		}
	}()

	<-ctx.Done()
	logger.Info("api_shutdown_start")
	shutdownCtx, cancel := context.WithTimeout(context.Background(), 15*time.Second)
	defer cancel()
	if err := srv.Shutdown(shutdownCtx); err != nil {
		logger.Error("shutdown_failed", zap.Error(err))
	}
	logger.Info("api_shutdown_done")
}

func newLogger(cfg *config.Config) (*zap.Logger, error) {
	if cfg.IsProd() {
		return zap.NewProduction()
	}
	return zap.NewDevelopment()
}

// maskURL 把 postgres://user:pass@host 里的密码换成 ***。
func maskURL(u string) string {
	if u == "" {
		return ""
	}
	if scheme, rest, ok := strings.Cut(u, "://"); ok {
		creds, host, ok := strings.Cut(rest, "@")
		if ok {
			if user, _, hasColon := strings.Cut(creds, ":"); hasColon {
				return scheme + "://" + user + ":***@" + host
			}
		}
	}
	return u
}